Create dedicated DatabaseCoordinator class
- intended to provide further databases files operations
- guaranties exclusive access to provided operations
