+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/Initializer.kt
@@ -1,15 +1,10 @@
 package cash.z.ecc.android.sdk
 
 import android.content.Context
-import android.os.Build
-import androidx.annotation.RequiresApi
+import cash.z.ecc.android.sdk.db.DatabaseCoordinator
 import cash.z.ecc.android.sdk.exception.InitializerException
 import cash.z.ecc.android.sdk.ext.ZcashSdk
-import cash.z.ecc.android.sdk.internal.AndroidApiVersion
-import cash.z.ecc.android.sdk.internal.SdkDispatchers
 import cash.z.ecc.android.sdk.internal.ext.getCacheDirSuspend
-import cash.z.ecc.android.sdk.internal.ext.getDatabasePathSuspend
-import cash.z.ecc.android.sdk.internal.ext.getNoBackupPathSuspend
 import cash.z.ecc.android.sdk.internal.model.Checkpoint
 import cash.z.ecc.android.sdk.internal.twig
 import cash.z.ecc.android.sdk.jni.RustBackend
@@ -18,11 +13,8 @@ import cash.z.ecc.android.sdk.tool.CheckpointTool
 import cash.z.ecc.android.sdk.tool.DerivationTool
 import cash.z.ecc.android.sdk.type.UnifiedViewingKey
 import cash.z.ecc.android.sdk.type.ZcashNetwork
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.withContext
 import java.io.File
-import java.util.*
 
 /**
  * Simplified Initializer focused on starting from a ViewingKey.
@@ -385,8 +377,8 @@ class Initializer private constructor(
             blockHeight: BlockHeight
         ): RustBackend {
             return RustBackend.init(
-                cacheDbPath(context, network, alias),
-                dataDbPath(context, network, alias),
+                DatabaseCoordinator.getInstance(context).cacheDbPath(network, alias),
+                DatabaseCoordinator.getInstance(context).dataDbPath(network, alias),
                 File(context.getCacheDirSuspend(), "params").absolutePath,
                 network,
                 blockHeight
@@ -411,318 +403,7 @@ class Initializer private constructor(
             appContext: Context,
             network: ZcashNetwork,
             alias: String
-        ): Boolean {
-            val cacheDeleted = deleteDb(cacheDbPath(appContext, network, alias))
-            val dataDeleted = deleteDb(dataDbPath(appContext, network, alias))
-            return dataDeleted || cacheDeleted
-        }
-
-        //
-        // Database Path Helpers
-        //
-
-        /**
-         * Returns the path to the Cache database that would correspond to the given alias
-         * and network attributes.
-         *
-         * @param appContext the application context
-         * @param network the network associated with the data in the cache database.
-         * @param alias the alias to convert into a database path
-         */
-        private suspend fun cacheDbPath(
-            appContext: Context,
-            network: ZcashNetwork,
-            alias: String
-        ): String {
-            val dbLocationsPair = prepareDbFiles(
-                appContext,
-                network,
-                alias,
-                ZcashSdk.DB_CACHE_NAME
-            )
-            return checkAndCopyDatabaseFiles(
-                dbLocationsPair.first,
-                dbLocationsPair.second
-            )
-        }
-
-        /**
-         * Returns the path to the Data database that would correspond to the given alias
-         * and network attributes.
-         *
-         * @param appContext the application context
-         * @param network the network associated with the data in the database.
-         * @param alias the alias to convert into a database path
-         */
-        private suspend fun dataDbPath(
-            appContext: Context,
-            network: ZcashNetwork,
-            alias: String
-        ): String {
-            val dbLocationsPair = prepareDbFiles(
-                appContext,
-                network,
-                alias,
-                ZcashSdk.DB_DATA_NAME
-            )
-            return checkAndCopyDatabaseFiles(
-                dbLocationsPair.first,
-                dbLocationsPair.second
-            )
-        }
-
-        /**
-         * Returns the path to the PendingTransaction database that would correspond to the given
-         * alias and network attributes. As the originally created file was called just
-         * PendingTransactions.db, we choose slightly different approach, but it also leads to
-         * original database files migration with additional renaming too.
-         *
-         * @param appContext the application context
-         * @param network the network associated with the data in the database.
-         * @param alias the alias to convert into a database path
-         */
-        suspend fun pendingTransactionsDbPath(
-            appContext: Context,
-            network: ZcashNetwork,
-            alias: String
-        ): String {
-            val legacyLocationDbFile = getDbFile(
-                null,
-                null,
-                ZcashSdk.DB_PENDING_TRANSACTIONS_NAME,
-                getLegacyDbParentDirPath(appContext)
-            )
-            val preferredLocationDbFile = if (AndroidApiVersion.isAtLeastL) {
-                getDbFile(
-                    network,
-                    alias,
-                    ZcashSdk.DB_PENDING_TRANSACTIONS_NAME,
-                    getDbParentDirPath(appContext)
-                )
-            } else {
-                null
-            }
-            return checkAndCopyDatabaseFiles(
-                legacyLocationDbFile,
-                preferredLocationDbFile
-            )
-        }
-
-        /**
-         * This helper function prepares a legacy (i.e. previously created) database file, as well
-         * as the preferred (i.e. newly created) file for subsequent use (and eventually move).
-         *
-         * Note: move of the database file is not performed for devices with Android SDK level lower
-         * than 21.
-         *
-         * @param appContext the application context
-         * @param network the network associated with the data in the database.
-         * @param alias the alias to convert into a database path
-         * @param databaseName the name of the new database file
-         */
-        private suspend fun prepareDbFiles(
-            appContext: Context,
-            network: ZcashNetwork,
-            alias: String,
-            databaseName: String
-        ): Pair<File, File?> {
-            val legacyLocationDbFile = getDbFile(
-                network,
-                alias,
-                databaseName,
-                getLegacyDbParentDirPath(appContext)
-            )
-            val preferredLocationDbFile = if (AndroidApiVersion.isAtLeastL) {
-                getDbFile(
-                    network,
-                    alias,
-                    databaseName,
-                    getDbParentDirPath(appContext)
-                )
-            } else {
-                null
-            }
-
-            return Pair(
-                legacyLocationDbFile,
-                preferredLocationDbFile
-            )
-        }
-
-        /**
-         * This function do actual database file copy or simply validate the file and return it.
-         * From the Android SDK level 21 it places database files into no_backup folder, as it does
-         * not allow automatic backup. On older APIs it places database files into databases folder,
-         * which allows automatic backup. It also copies database files between these two folders,
-         * if older folder usage is detected.
-         *
-         * @param legacyLocationDbFile the previously used file location
-         * @param preferredLocationDbFile the newly used file location
-         */
-        private suspend fun checkAndCopyDatabaseFiles(
-            legacyLocationDbFile: File,
-            preferredLocationDbFile: File?
-        ): String = withContext(Dispatchers.IO) {
-            var preferredDbFile = preferredLocationDbFile
-
-            // we run the copy action only on Android SDK 21 and higher
-            val resultDbFile = if (preferredDbFile != null) {
-
-                // check if the copy wasn't already performed and if it's needed
-                if (!preferredDbFile.exists() && legacyLocationDbFile.exists()) {
-                    // We check the copy operation result and fallback to the legacy file, if
-                    // anything went wrong.
-                    if (!copyDatabaseFile(legacyLocationDbFile, preferredDbFile)) {
-                        preferredDbFile = legacyLocationDbFile
-                    }
-                }
-                preferredDbFile
-            } else {
-                legacyLocationDbFile
-            }
-            resultDbFile.absolutePath
-        }
-
-        /**
-         * The purpose of this function is to copy database files between the old location (given by
-         * the legacyLocationDbFile parameter) and the new location (given by preferredLocationDbFile).
-         * The actual copy operation is performed with the renameTo function, which simply renames
-         * a file path and persists the metadata information. The mechanism deals with the additional
-         * database files -journal and -wal too, if they exist.
-         *
-         * @param legacyLocationDbFile the previously used file location (rename from)
-         * @param preferredLocationDbFile the newly used file location (rename to)
-         */
-        private fun copyDatabaseFile(legacyLocationDbFile: File, preferredLocationDbFile: File): Boolean {
-            val filesToBeRenamed = LinkedList<Pair<File, File>>()
-            filesToBeRenamed.add(Pair(legacyLocationDbFile, preferredLocationDbFile))
-
-            // add journal database file, if exists
-            val journalSuffixedDbFile = File(
-                legacyLocationDbFile.absolutePath + ZcashSdk.DATABASE_FILE_JOURNAL_SUFFIX
-            )
-            if (journalSuffixedDbFile.exists()) {
-                filesToBeRenamed.add(
-                    Pair(
-                        journalSuffixedDbFile,
-                        File(preferredLocationDbFile.absolutePath + ZcashSdk.DATABASE_FILE_JOURNAL_SUFFIX)
-                    )
-                )
-            }
-
-            // add wal database file, if exists
-            val walSuffixedDbFile = File(
-                legacyLocationDbFile.absolutePath + ZcashSdk.DATABASE_FILE_WAL_SUFFIX
-            )
-            if (walSuffixedDbFile.exists()) {
-                filesToBeRenamed.add(
-                    Pair(
-                        walSuffixedDbFile,
-                        File(preferredLocationDbFile.absolutePath + ZcashSdk.DATABASE_FILE_WAL_SUFFIX)
-                    )
-                )
-            }
-
-            return runCatching {
-                return@runCatching filesToBeRenamed.all {
-                    it.first.renameTo(it.second)
-                }
-            }.onFailure {
-                twig("Failed while renaming database files with: $it")
-            }.getOrDefault(false)
-        }
-
-        /**
-         * This function returns previously used database folder path (i.e. databases). The databases
-         * folder is deprecated now, as it allows automatic data backup, which is not permitted for
-         * our database files.
-         *
-         * @param appContext the application context
-         */
-        private suspend fun getLegacyDbParentDirPath(appContext: Context): String {
-            return appContext.getDatabasePathSuspend("unused.db").parentFile?.absolutePath
-                ?: throw InitializerException.DatabasePathException
-        }
-
-        /**
-         * This function returns a newly used database folder path (i.e. no_backup). The databases
-         * folder is deprecated now and we use no_backup to avoid automatic backup of large possibly
-         * sensitive data.
-         *
-         * @param appContext the application context
-         */
-        @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
-        private suspend fun getDbParentDirPath(appContext: Context): String {
-            return appContext.getNoBackupPathSuspend().absolutePath
-                ?: throw InitializerException.DatabasePathException
-        }
-
-        /**
-         * Simple helper function, which prepares a database file object by input parameters.
-         *
-         * @param network the network associated with the data in the database.
-         * @param alias the alias to convert into a database path
-         * @param dbFileName the name of the new database file
-         * @param parentDir the name of the parent directory, in which the file should be placed
-         */
-        private fun getDbFile(
-            network: ZcashNetwork?,
-            alias: String?,
-            dbFileName: String,
-            parentDir: String
-        ): File {
-            val aliasPrefix = if (alias == null) {
-                ""
-            } else if (alias.endsWith('_')) {
-                alias
-            } else {
-                "${alias}_"
-            }
-
-            val networkPrefix = network?.networkName ?: ""
-
-            return if (aliasPrefix.isNotEmpty()) {
-                File(parentDir, "$aliasPrefix${networkPrefix}_$dbFileName")
-            } else {
-                File(parentDir, dbFileName)
-            }
-        }
-
-        /**
-         * Delete a database and it's potential journal file at the given path.
-         *
-         * The rollback journal file is a temporary file used to implement atomic commit and
-         * rollback capabilities in SQLite.
-         *
-         * @param filePath the path of the db to erase.
-         * @return true when a file exists at the given path and was deleted.
-         */
-        private suspend fun deleteDb(filePath: String): Boolean {
-            // just try the journal file. Doesn't matter if it's not there.
-            delete("$filePath-journal")
-
-            return delete(filePath)
-        }
-
-        /**
-         * Delete the file at the given path.
-         *
-         * @param filePath the path of the file to erase.
-         * @return true when a file exists at the given path and was deleted.
-         */
-        private suspend fun delete(filePath: String): Boolean {
-            return File(filePath).let {
-                withContext(SdkDispatchers.DATABASE_IO) {
-                    if (it.exists()) {
-                        twig("Deleting ${it.name}!")
-                        it.delete()
-                        true
-                    } else {
-                        false
-                    }
-                }
-            }
-        }
+        ): Boolean = DatabaseCoordinator.getInstance(appContext).deleteDatabases(network, alias)
     }
 }
